requirements.txt
python-telegram-bot==20.7
Pillow==10.2.0
python-dotenv==1.0.1
requests==2.31.0
google-cloud-vision==3.5.0 
flask[async] 
pytesseract 
rapidfuzz 